put into form as described by ticket; most reporting to TS
