+++ backend/package.json
@@ -30,7 +30,6 @@
     "express-flash": "0.0.2",
     "express-session": "^1.17.2",
     "geckodriver": "^3.0.1",
-    "lodash": "^4.17.21",
     "moment": "^2.29.4",
     "mongodb": "^3.7.3",
     "node-fetch": "^2.6.7",
