+++ backend/src/serverRouter/jiraRouter.js
@@ -2,7 +2,7 @@ const express = require('express');
 const cors = require('cors');
 const bodyParser = require('body-parser');
 const fetch = require('node-fetch');
-const userHelper = require('../../dist/user');
+const userHelper = require('../../dist/userManagement');
 
 const router = express.Router();
